Solution for 2020 Day 2 Part 1
This one is also pretty straightforward. The most complex part is
parsing each line of the input to get four separate tokens: the
password, the needed character, and the lower and upper count limits
for that character. If I were doing this in Python, I'd use a regex
with capture groups and extract those groups from the match object.
In Go, I'm honestly not sure how to do that. Luckily, the format is
easy to parse with just strings.Split().

I'd also consider doing this with a separate goroutine doing all of
the actual processing, but it doesn't actually get us any efficiency
gains (unless I spun up multiple goroutines to process in parallel).