Solution for 2020 Day 2 Part 2
I was briefly thrown off by the fact that indexing into a string
returns a `[]byte`, rather than a `rune`, and got halfway through
changing my `policy` struct to hold a rune instead of a string for
the character to match. But then I realized that I could just convert
the `[]byte`s to strings before comparing and call it a day.

Also slightly sad Go doesn't have a logical xor operator for booleans,
but != is good enough.