Solution for 2020 Day 3 Part 2
Yep, this is basically what I expected (although I expected to have
to return the minium of the numbers, not their product).

I factored the heavy lifting out into a helper, so that I could call
it multiple times for this solution. But I did have to change my
approach from "iterate over the rows of the map" to "iterate over
all indices generated by the slope". This results in code which is
marginally faster, but definitely more flexible and arguably easier
to reason about.