Update Day 2 solution to use regexps
The way the go documentation talks about regexps is rather confusing.
I read the docs twice and still wasn't able to realize that there is
in fact an implementation of capture groups. Once I went back and
figured that out, rewriting this one using capture groups instead of
string splitting was a no-brainer.